MachineLearningPlatforms/Kubeflow/ml-components/women_clothing_ecom_reviews/pipeline/components/data_collection/requirements.txt
pyyaml==5.4.1
kaggle==1.5.10
pandas==1.2.3
imbalanced-learn==0.8.0
scikit-learn==0.24.1